Renamed sample to FeatureLayerGeoPackage for consistency with other sample names
