+++ src/Android/Xamarin.Android/Samples/Data/FeatureLayerGeoPackage/FeatureLayerGeoPackage.jpg
