+++ src/Forms/Shared/Forms.projitems
@@ -36,10 +36,10 @@
       <SubType>Code</SubType>
       <DependentUpon>AddShapefile.xaml</DependentUpon>
     </Compile>
-    <Compile Include="$(MSBuildThisFileDirectory)Samples\Data\AddGeoPackageFeatureTable\AddGeoPackageFeatureTable.xaml.cs">
+    <Compile Include="$(MSBuildThisFileDirectory)Samples\Data\FeatureLayerGeoPackage\FeatureLayerGeoPackage.xaml.cs">
       <SubType>Code</SubType>
-      <DependentUpon>AddGeoPackageFeatureTable.xaml</DependentUpon>
-    </Compile>    
+      <DependentUpon>FeatureLayerGeoPackage.xaml</DependentUpon>
+    </Compile>
     <Compile Include="$(MSBuildThisFileDirectory)Samples\Data\ServiceFeatureTableManualCache\ServiceFeatureTableManualCache.xaml.cs">
       <SubType>Code</SubType>
       <DependentUpon>ServiceFeatureTableManualCache.xaml</DependentUpon>
@@ -276,11 +276,11 @@
     <EmbeddedResource Include="$(MSBuildThisFileDirectory)Samples\Data\AddShapefile\AddShapefile.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
-    <EmbeddedResource Include="$(MSBuildThisFileDirectory)Samples\Data\AddGeoPackageFeatureTable\metadata.json" />
-    <EmbeddedResource Include="$(MSBuildThisFileDirectory)Samples\Data\AddGeoPackageFeatureTable\AddGeoPackageFeatureTable.jpg" />
-    <EmbeddedResource Include="$(MSBuildThisFileDirectory)Samples\Data\AddGeoPackageFeatureTable\AddGeoPackageFeatureTable.xaml">
+    <EmbeddedResource Include="$(MSBuildThisFileDirectory)Samples\Data\FeatureLayerGeoPackage\metadata.json" />
+    <EmbeddedResource Include="$(MSBuildThisFileDirectory)Samples\Data\FeatureLayerGeoPackage\FeatureLayerGeoPackage.jpg" />
+    <EmbeddedResource Include="$(MSBuildThisFileDirectory)Samples\Data\FeatureLayerGeoPackage\FeatureLayerGeoPackage.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
-    </EmbeddedResource>    
+    </EmbeddedResource>
     <EmbeddedResource Include="$(MSBuildThisFileDirectory)Samples\Data\ServiceFeatureTableManualCache\metadata.json" />
     <EmbeddedResource Include="$(MSBuildThisFileDirectory)Samples\Data\ServiceFeatureTableManualCache\ServiceFeatureTableManualCache.jpg">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
