+++ src/Forms/Shared/Samples/Data/FeatureLayerGeoPackage/FeatureLayerGeoPackage.xaml
@@ -3,6 +3,6 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"             
              xmlns:esriUI="clr-namespace:Esri.ArcGISRuntime.Xamarin.Forms;assembly=Esri.ArcGISRuntime.Xamarin.Forms"
              xmlns:mapping="clr-namespace:Esri.ArcGISRuntime.Mapping;assembly=Esri.ArcGISRuntime" 
-             x:Class="ArcGISRuntimeXamarin.Samples.AddGeoPackageFeatureTable.AddGeoPackageFeatureTable">
+             x:Class="ArcGISRuntimeXamarin.Samples.FeatureLayerGeoPackage.FeatureLayerGeoPackage">
     <esriUI:MapView x:Name="MyMapView"/>
 </ContentPage>
