+++ src/UWP/ArcGISRuntime.UWP.Samples/ArcGISRuntime.UWP.Samples.CSharp.csproj
@@ -85,12 +85,12 @@
   </PropertyGroup>
   <ItemGroup>
     <!-- A reference to the entire .Net Framework and Windows SDK are automatically included -->
-    <Content Include="Samples\Data\AddGeoPackageFeatureTable\AddGeoPackageFeatureTable.jpg">
+    <Content Include="Samples\Data\FeatureLayerGeoPackage\FeatureLayerGeoPackage.jpg">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="Samples\Data\AddShapefile\AddShapefile.jpg">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>    
+    </Content>
     <Content Include="Samples\Data\StatisticalQuery\StatisticalQuery.jpg">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
@@ -459,13 +459,13 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
-    <Content Include="Samples\Data\AddGeoPackageFeatureTable\metadata.json">
+    <Content Include="Samples\Data\FeatureLayerGeoPackage\metadata.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
-    <Page Include="Samples\Data\AddGeoPackageFeatureTable\AddGeoPackageFeatureTable.xaml">
+    <Page Include="Samples\Data\FeatureLayerGeoPackage\FeatureLayerGeoPackage.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
-    </Page>    
+    </Page>
     <Page Include="Samples\Layers\KmlLayerFile\KmlLayerFile.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
@@ -492,12 +492,12 @@
     <Compile Include="Samples\Data\GenerateGeodatabase\GenerateGeodatabase.xaml.cs">
       <DependentUpon>GenerateGeodatabase.xaml</DependentUpon>
     </Compile>
-    <Compile Include="Samples\Data\AddGeoPackageFeatureTable\AddGeoPackageFeatureTable.xaml.cs">
-      <DependentUpon>AddGeoPackageFeatureTable.xaml</DependentUpon>
+    <Compile Include="Samples\Data\FeatureLayerGeoPackage\FeatureLayerGeoPackage.xaml.cs">
+      <DependentUpon>FeatureLayerGeoPackage.xaml</DependentUpon>
     </Compile>
     <Compile Include="Samples\Data\AddShapefile\AddShapefile.xaml.cs">
       <DependentUpon>AddShapefile.xaml</DependentUpon>
-    </Compile>    
+    </Compile>
     <Compile Include="Samples\Data\ServiceFeatureTableCache\ServiceFeatureTableCache.xaml.cs">
       <DependentUpon>ServiceFeatureTableCache.xaml</DependentUpon>
     </Compile>
