+++ src/UWP/ArcGISRuntime.UWP.Samples/Samples/Data/FeatureLayerGeoPackage/FeatureLayerGeoPackage.jpg
