+++ src/UWP/ArcGISRuntime.UWP.Samples/Samples/Data/FeatureLayerGeoPackage/FeatureLayerGeoPackage.xaml
@@ -1,5 +1,5 @@
 ﻿<UserControl 
-    x:Class="ArcGISRuntime.UWP.Samples.AddGeoPackageFeatureTable.AddGeoPackageFeatureTable"
+    x:Class="ArcGISRuntime.UWP.Samples.FeatureLayerGeoPackage.FeatureLayerGeoPackage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:esriUI="using:Esri.ArcGISRuntime.UI.Controls" >
