+++ src/UWP/ArcGISRuntime.UWP.Samples/Samples/Data/FeatureLayerGeoPackage/metadata.json
@@ -1,6 +1,6 @@
 {
-  "Name": "Add GeoPackage feature table",
-  "SampleName": "AddGeoPackageFeatureTable",
+  "Name": "Feature layer (GeoPackage)",
+  "SampleName": "FeatureLayerGeoPackage",
   "Description": "This sample demonstrates how to open a GeoPackage and show a GeoPackage feature table in a feature layer.",
   "Instructions": "The GeoPackage will be downloaded from an ArcGIS Online portal automatically.",
   "Type": 0,
@@ -8,7 +8,7 @@
   "RequiresOfflineData": true,
   "DataItemIds": [ "68ec42517cdd439e81b036210483e8e7" ],
   "RequiresLocalServer": false,
-  "Image": "AddGeoPackageFeatureTable.jpg",
+  "Image": "FeatureLayerGeoPackage.jpg",
   "Link": "",
   "TypeLink": [ "T:Esri.ArcGISRuntime.Data.GeoPackage", "T:Esri.ArcGISRuntime.Data.GeoPackageFeatureTable", "P:Esri.ArcGISRuntime.Data.GeoPackage.GeoPackageFeatureTables", "M:Esri.ArcGISRuntime.Data.GeoPackage.OpenAsync(System.String)" ]
 }
