+++ src/WPF/ArcGISRuntime.WPF.Samples/ArcGISRuntime.WPF.Samples.CSharp.csproj
@@ -56,8 +56,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Samples\Data\AddGeoPackageFeatureTable\AddGeoPackageFeatureTable.xaml.cs">
-      <DependentUpon>AddGeoPackageFeatureTable.xaml</DependentUpon>
+    <Compile Include="Samples\Data\FeatureLayerGeoPackage\FeatureLayerGeoPackage.xaml.cs">
+      <DependentUpon>FeatureLayerGeoPackage.xaml</DependentUpon>
     </Compile>
     <Compile Include="Samples\Data\AddShapefile\AddShapefile.xaml.cs">
       <DependentUpon>AddShapefile.xaml</DependentUpon>
@@ -247,12 +247,12 @@
     <Content Include="groups.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
-    <Content Include="Samples\Data\AddGeoPackageFeatureTable\AddGeoPackageFeatureTable.jpg">
+    <Content Include="Samples\Data\FeatureLayerGeoPackage\FeatureLayerGeoPackage.jpg">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="Samples\Data\AddShapefile\AddShapefile.jpg">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>    
+    </Content>
     <Content Include="Samples\Data\StatisticalQuery\StatisticalQuery.jpg">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
@@ -442,14 +442,14 @@
     </Content>
   </ItemGroup>
   <ItemGroup>
-    <Page Include="Samples\Data\AddGeoPackageFeatureTable\AddGeoPackageFeatureTable.xaml">
+    <Page Include="Samples\Data\FeatureLayerGeoPackage\FeatureLayerGeoPackage.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
     <Page Include="Samples\Data\AddShapefile\AddShapefile.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
-    </Page>    
+    </Page>
     <Page Include="Samples\Data\StatisticalQuery\StatisticalQuery.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
@@ -928,7 +928,7 @@
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
-    <Content Include="Samples\Data\AddGeoPackageFeatureTable\metadata.json">
+    <Content Include="Samples\Data\FeatureLayerGeoPackage\metadata.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="Samples\Data\AddShapefile\metadata.json">
