+++ src/WPF/ArcGISRuntime.WPF.Samples/Samples/Data/FeatureLayerGeoPackage/FeatureLayerGeoPackage.jpg
