+++ src/iOS/Xamarin.iOS/groups.json
@@ -244,8 +244,8 @@
               "Path": "Samples/Data/AddShapefile"
             },
             {
-              "SampleName": "AddGeoPackageFeatureTable",
-              "Path": "Samples/Data/AddGeoPackageFeatureTable"
+              "SampleName": "FeatureLayerGeoPackage",
+              "Path": "Samples/Data/FeatureLayerGeoPackage"
             }
           ]
         }
